Add basic struct support to schema builder
Can handle nested and recursive structs.

It can:
 * taking "json:name" into account
 * handle nested structs
 * handle recursive struct definitions

When providing a struct, the struct is recursively walked through and
for all struct types and primitive types the properties and schemas are
generated. As a result of the `SchemaBuilder#Build(struct)` method you
get a reference schema to the provided struct as the first result value,
the second return value contains all schemas which the builder found.
They can be placed in the `definitions` section of a swagger file.